+++ tests/test_runner.py
@@ -18,8 +18,8 @@ import tests.helper as test_module
 from hera.shared import GlobalConfig
 from hera.shared._pydantic import _PYDANTIC_VERSION
 from hera.shared.serialization import serialize
+from hera.workflows._runner.util import _run, _runner
 from hera.workflows.io.v1 import RunnerOutput
-from hera.workflows.runner import _run, _runner
 from hera.workflows.script import RunnerScriptConstructor
 
 
@@ -614,8 +614,8 @@ def test_script_annotations_unknown_type(global_config_fixture: GlobalConfig):
         [{"name": "a_number", "value": 123}],
     ],
 )
-@patch("hera.workflows.runner._runner")
-@patch("hera.workflows.runner._parse_args")
+@patch("hera.workflows._runner.util._runner")
+@patch("hera.workflows._runner.util._parse_args")
 def test_run(mock_parse_args, mock_runner, kwargs_list, tmp_path: Path):
     # GIVEN
     file_path = Path(tmp_path / "test_params")
@@ -633,8 +633,8 @@ def test_run(mock_parse_args, mock_runner, kwargs_list, tmp_path: Path):
     mock_runner.assert_called_once_with("my_entrypoint", kwargs_list)
 
 
-@patch("hera.workflows.runner._runner")
-@patch("hera.workflows.runner._parse_args")
+@patch("hera.workflows._runner.util._runner")
+@patch("hera.workflows._runner.util._parse_args")
 def test_run_empty_file(mock_parse_args, mock_runner, tmp_path: Path):
     # GIVEN
     file_path = Path(tmp_path / "test_params")
@@ -652,8 +652,8 @@ def test_run_empty_file(mock_parse_args, mock_runner, tmp_path: Path):
     mock_runner.assert_called_once_with("my_entrypoint", [])
 
 
-@patch("hera.workflows.runner._runner")
-@patch("hera.workflows.runner._parse_args")
+@patch("hera.workflows._runner.util._runner")
+@patch("hera.workflows._runner.util._parse_args")
 def test_run_null_string(mock_parse_args, mock_runner, tmp_path: Path):
     # GIVEN
     file_path = Path(tmp_path / "test_params")
@@ -935,7 +935,7 @@ def test_runner_pydantic_output_with_exit_code(
         ),
     ],
 )
-@patch("hera.workflows.runner._parse_args")
+@patch("hera.workflows._runner.util._parse_args")
 def test_run_pydantic_output_with_exit_code(
     mock_parse_args,
     entrypoint,
